SoundManager: renamed to SoundManagerBase
